+++ subpackages/cli/source/odood/cli/commands/odoo.d
@@ -9,6 +9,7 @@ private import odood.cli.core: OdoodCommand;
 private import odood.lib.project: Project;
 
 
+// TODO: May be move all this to dev-tools
 class CommandOdooShell: OdoodCommand {
     this() {
         super("shell", "Odoo-related utility commands.");
@@ -27,10 +28,33 @@ class CommandOdooShell: OdoodCommand {
 }
 
 
+class CommandOdooRecomputeField: OdoodCommand {
+    this() {
+        super("recompute", "Odoo-related utility commands.");
+        this.add(
+            new Option(
+                "f", "field", "Name of field to recompute."
+            ).repeating);
+        this.add(
+            new Argument(
+                "dbname", "Name of database to recompute fields for").required);
+        this.add(
+            new Argument(
+                "model", "Name of model to recompute fields for").required);
+
+    }
+
+    public override void execute(ProgramArgs args) {
+        Project.loadProject.lodoo.recomputeField(args.arg("dbname"), args.arg("model"), args.options("field"));
+    }
+}
+
+
 class CommandOdoo: OdoodCommand {
     this() {
         super("odoo", "Odoo-related utility commands.");
         this.add(new CommandOdooShell());
+        this.add(new CommandOdooRecomputeField());
     }
 }