+++ test-data/test_addon/__manifest__.py
@@ -1,0 +1,8 @@
+{
+    "name": "Test addon",
+    "version": "18.0.1.2.0",
+    "depends": [
+        "base",
+        "web",
+    ],
+}